* Ships now look smoother when traveling - the pixel-putting was ugly
git-svn-id: https://svn.code.sf.net/p/qonk/code/trunk@5 917c0414-312c-0410-a5b5-8fb0c60c5ce6

+++ ships.h
@@ -53,7 +53,7 @@ public:
   Player* getOwner() const;
   Planet* getPlanet() const;
 
-  int getDirection() const;
+  double getDirection() const;
 };
 
 class Ships : public list< Ship* > {
